i2c-designware-core: reset the controller if no reset/abort
In some cases the controller does not set any reset or abort interrupt,
gets crazy and generates a lot of interrupts, leading to soft lockup.
If it happens, mask all the interrupts to let the transaction timeout,
leading to a reset of the controller to go back in a normal state.

Change-Id: Ic3e8829d3ac2e70998daa6badd4c4fc35d3f17b6
Tracked-On: https://jira01.devtools.intel.com/browse/AW-4892
Signed-off-by: Simon Dubray <simonx.dubray@intel.com>
Reviewed-on: https://android.intel.com/580379
Reviewed-by: jenkins_ndg <jenkins_ndg@intel.com>
Reviewed-by: Nassiet, GaelleX <gaellex.nassiet@intel.com>
Reviewed-by: Hu, Bingquan <bingquan.hu@intel.com>
Reviewed-by: Whitfield, MichaelX <michaelx.whitfield@intel.com>
Reviewed-by: Zaghdoud, WalidX <walidx.zaghdoud@intel.com>
Reviewed-by: Tasayco Loarte, VictorX <victorx.tasayco.loarte@intel.com>
Reviewed-by: Choudhary, ShriramX <shriramx.choudhary@intel.com>