ANDROID: binder: fix OOB write in __binder_update_page_range
commit e12f1f1c65a9 ("binder: always allocate/map first
BINDER_MIN_ALLOC pages") introduced a vulnerability where
the alloc->pages array can be accessed beyond the allocated
elements. For this to occur, the passed-in VMA range must
be < BINDER_MIN_ALLOC pages (which cannot happen for normal
use cases). BINDER_MIN_ALLOC is set to 6 pages, so mmap'ing
1 page results in OOB accesses.

The fix is to return an error if binder_mmap is called with
a VMA size less than BINDER_MIN_ALLOC.

Bug: 64216036
Test: verified fixed with submitter's test program
Change-Id: I4b6dbf3f466fea46e587d6a33f724ee4fe7f1368
Tracked-On: https://jira01.devtools.intel.com/browse/AW-7591
Signed-off-by: Todd Kjos <tkjos@google.com>
Signed-off-by: Victor Tasayco Loarte <victorx.tasayco.loarte@intel.com>
Reviewed-on: https://android.intel.com/611663
Reviewed-by: Dubray, SimonX <simonx.dubray@intel.com>
Reviewed-by: Louis, FabienX <fabienx.louis@intel.com>
Reviewed-by: Zaghdoud, WalidX <walidx.zaghdoud@intel.com>
Reviewed-by: Akue, LoicX <loicx.akue@intel.com>