arch/x86/configs: cleanup SND_* support
- disable MFD_WM8994
This is for Wolfson Microelectronics WM8350 with I2C, which is an
audiocodec.  We do not have this audio-codec.

- disable CONFIG_SND_USB
We don't have USB audio support

- disable CONFIG_SND_SEQUENCER
SND_SEQUENCER is for MIDI sequencing.
This is not something supported on our watch.

- set max sound-cards to 4
Default of 32 is optimistic. We currently have only two sound-cards.
Take some margin and set default to 4.

Change-Id: If4c56eb079a23918224cf28da3dc00a4bd2504e4
Tracked-On: https://jira01.devtools.intel.com/browse/AW-7638
Signed-off-by: Mattijs Korpershoek <mattijsx.korpershoek@intel.com>
Reviewed-on: https://android.intel.com/616454
Reviewed-by: Dubray, SimonX <simonx.dubray@intel.com>
Reviewed-by: jenkins_ndg <jenkins_ndg@intel.com>
Reviewed-by: Ledentec, AlexandreX <alexandrex.ledentec@intel.com>
Reviewed-by: Khaldi, RiadhX <riadhx.khaldi@intel.com>
Reviewed-by: Akue, LoicX <loicx.akue@intel.com>