intel_media/drm: add support of non-speculative fences
In HWC2.0, retire/release fences are non-speculative. A retire fence
for a frame N should signal as soon as the frame is pushed to the display.

This patch aligns the way we handle retire fences, at DC server, to HWC2.0
requirements. We're expected to release the display buffer as soon as the
flip takes effect on screen.

Change-Id: Ia08d0ec4507aae0f68807845c46fc52430de2717
Tracked-On: https://jira01.devtools.intel.com/browse/AW-7929
Signed-off-by: Zaghdoud Walidx <walidx.zaghdoud@intel.com>
Reviewed-on: https://android.intel.com/638048
Reviewed-by: jenkins_ndg <jenkins_ndg@intel.com>
Reviewed-by: Dubray, SimonX <simonx.dubray@intel.com>
Reviewed-by: Ledentec, AlexandreX <alexandrex.ledentec@intel.com>
Reviewed-by: Viel, ClementX <clementx.viel@intel.com>
Reviewed-by: Akue, LoicX <loicx.akue@intel.com>
Reviewed-by: Tasayco Loarte, VictorX <victorx.tasayco.loarte@intel.com>
Reviewed-on: https://android.intel.com/644289
Tested-by: Tasayco Loarte, VictorX <victorx.tasayco.loarte@intel.com>